// src/client/types.gen.ts
export type IdentityUserDto = {
	readonly extraProperties?: {
		[key: string]: unknown;
	} | null;
	id?: string;
	creationTime?: string;
	creatorId?: string | null;
	lastModificationTime?: string | null;
	lastModifierId?: string | null;
	isDeleted?: boolean;
	deleterId?: string | null;
	deletionTime?: string | null;
	tenantId?: string | null;
	userName?: string | null;
	name?: string | null;
	surname?: string | null;
	email?: string | null;
	emailConfirmed?: boolean;
	phoneNumber?: string | null;
	phoneNumberConfirmed?: boolean;
	isActive?: boolean;
	lockoutEnabled?: boolean;
	accessFailedCount?: number;
	lockoutEnd?: string | null;
	concurrencyStamp?: string | null;
	entityVersion?: number;
	lastPasswordChangeTime?: string | null;
};
// ...
export type PermissionGrantInfoDto = {
	name?: string | null;
	displayName?: string | null;
	parentName?: string | null;
	isGranted?: boolean;
	allowedProviders?: Array<string> | null;
	grantedProviders?: Array<ProviderInfoDto> | null;
};

export type PermissionGroupDto = {
	name?: string | null;
	displayName?: string | null;
	displayNameKey?: string | null;
	displayNameResource?: string | null;
	permissions?: Array<PermissionGrantInfoDto> | null;
};
// ...
export type ProviderInfoDto = {
	providerName?: string | null;
	providerKey?: string | null;
};